# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "contact": {
            "name": "Krisna Cipta Raksa",
            "email": "kciptaraksa@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/order": {
            "get": {
                "tags": [
                    "Get All Order"
                ],
                "summary": "Fetch order data (using Firebase Realtime Database)",
                "description": "Fetch all data order.",
                "operationId": "fetchorder",
                "responses": {
                    "200": {
                        "description": "Fetch data success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_order_success": {
                                        "summary": "Response while fetch data successfully",
                                        "value": {
                                            "status": "success",
                                            "message": "Data fetched successfully.",
                                            "data": "Order data"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to fetch data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_order_failed": {
                                        "summary": "Response while fetch data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Failed to fetch data",
                                            "data": ""
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Request without Token",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "example": "Unauthorized."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order/store": {
            "post": {
                "tags": [
                    "Store Order"
                ],
                "summary": "Store order data (using Firebase Realtime Database)",
                "description": "Store order data.",
                "operationId": "storeorder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer": {
                                        "type": "string"
                                    },
                                    "product": {
                                        "type": "string"
                                    },
                                    "qty": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "customer": "Amir Khan",
                                    "product": "Bardi Smart Lamp",
                                    "qty": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_store_data_Order_success": {
                                        "summary": "Response while store data successfully",
                                        "value": {
                                            "status": "success",
                                            "message": "Order created successfully.",
                                            "data": "Order data inserted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to store data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_store_data_order_failed": {
                                        "summary": "Response while store data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Order created failed.",
                                            "data": ""
                                        }
                                    },
                                    "response_while_store_data_order_failed_because_mandatory_field_empty": {
                                        "summary": "Response while store data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Mandatory field cannot be empty",
                                            "data": "Validation Information"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Request without Token",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "example": "Unauthorized."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order/{id}/show": {
            "get": {
                "tags": [
                    "Get Order By ID"
                ],
                "summary": "Fetch order data by ID (using Firebase Realtime Database)",
                "description": "Fetch order data by ID.",
                "operationId": "fetchOrderbyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetch data success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_order_byid_success": {
                                        "summary": "Response while fetch data successfully",
                                        "value": {
                                            "status": "success",
                                            "message": "Data fetched successfully.",
                                            "data": "Order data"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to fetch data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_order_byid_failed": {
                                        "summary": "Response while fetch data failed",
                                        "value": {
                                            "status": "error",
                                            "message": "Failed to fetch data",
                                            "data": ""
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Request without Token",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "example": "Unauthorized."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Failed to fetch data because ID not given",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "response_while_fetch_data_order_byid_failed_because_no_id": {
                                        "summary": "Response while fetch data failed because ID not given",
                                        "value": {
                                            "status": "error",
                                            "message": "Page Not Found",
                                            "data": ""
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get authentication token",
                "name": "apiAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}